requirements.txt
pandas
numpy
python-telegram-bot
python-telegram-bot[job-queue]
matplotlib
PyGithub
-e .